pr comment fixes
