+++ src/Domain.Linnapps/Domain.Linnapps.csproj
@@ -10,7 +10,6 @@
     <PackageReference Include="Linn.Common.Configuration" Version="2.0.0" />
     <PackageReference Include="Linn.Common.Reporting" Version="1.0.0" />
     <PackageReference Include="Linn.Common.Serialization" Version="2.1.0" />
-    <PackageReference Include="Oracle.ManagedDataAccess.Core" Version="2.18.3" />
   </ItemGroup>
 
 </Project>
