+++ src/Domain.Linnapps/RemoteServices/IStockTriggerLevelReportService.cs
@@ -1,0 +1,11 @@
+﻿namespace Linn.Products.Domain.Linnapps.RemoteServices
+{
+    using System.Data;
+
+    public interface IStockTriggerLevelReportService
+    {
+        DataTable GetStockTriggerLevelsForPartAtLocation(int locationId, string partNumber);
+
+        DataTable GetPartDataAtLocation(int locationId);
+    }
+}
