+++ src/Facade/Services/StockTriggerLevelReportService.cs
@@ -1,29 +1,27 @@
 ﻿namespace Linn.Products.Facade.Services
 {
-    using System.Collections.Generic;
-
     using Linn.Common.Facade;
     using Linn.Common.Reporting.Models;
-    using Linn.Products.Domain.Reports;
+    using Linn.Products.Domain.Linnapps.Reports;
 
-    public class StockTriggerLevelReportService : IStockTriggerLevelReportService
+    public class StockTriggerLevelReportService : IStockTriggerLevelsService
     {
-        private readonly Domain.Linnapps.Reports.IStockTriggerLevelReportService stockTriggerLevelReportService;
+        private readonly IStockTriggerLevelReportService stockTriggerLevelsService;
 
-        public StockTriggerLevelReportService(Domain.Linnapps.Reports.IStockTriggerLevelReportService stockTriggerLevelReportService)
+        public StockTriggerLevelReportService(Domain.Linnapps.Reports.IStockTriggerLevelReportService stockTriggerLevelsService)
         {
-            this.stockTriggerLevelReportService = stockTriggerLevelReportService;
+            this.stockTriggerLevelsService = stockTriggerLevelsService;
         }
 
         public IResult<ResultsModel> GetStockTriggerLevelsForPartAtLocation(int locationId, string partNumber)
         {
-            var results = this.stockTriggerLevelReportService.GetStockTriggerLevelReportForPartAtLocation(locationId, partNumber);
+            var results = this.stockTriggerLevelsService.GetStockTriggerLevelReportForPartAtLocation(locationId, partNumber);
             return new SuccessResult<ResultsModel>(results);
         }
 
         public IResult<ResultsModel> GetPartDataAtLocation(int locationId)
         {
-            var results = this.stockTriggerLevelReportService.GetPartDataAtLocation(locationId);
+            var results = this.stockTriggerLevelsService.GetPartDataAtLocation(locationId);
             return new SuccessResult<ResultsModel>(results);
         }
     }
