+++ src/IoC/ServiceModule.cs
@@ -13,6 +13,8 @@
     using Linn.Products.Proxy;
     using Linn.Products.Resources;
 
+    using IStockTriggerLevelReportService = Linn.Products.Domain.Linnapps.RemoteServices.IStockTriggerLevelReportService;
+
     public class ServiceModule : Module
     {
         protected override void Load(ContainerBuilder builder)
@@ -32,7 +34,7 @@
             builder.RegisterType<SaCoreTypeService>().As<IFacadeService<SaCoreType, int, SaCoreTypeResource>>(); 
             builder.RegisterType<SernosConfigService>().As<IFacadeService<SernosConfig, string, SernosConfigResource>>();
             builder.RegisterType<TypeOfSaleService>().As<IFacadeService<TypeOfSale, string, TypeOfSaleResource>>();
-            builder.RegisterType<Facade.Services.StockTriggerLevelReportService>().As<Facade.Services.IStockTriggerLevelReportService>();
+            builder.RegisterType<Facade.Services.StockTriggerLevelReportService>().As<Facade.Services.IStockTriggerLevelsService>();
  
             // rest client proxies
             builder.RegisterType<RestClient>().As<IRestClient>();
@@ -40,6 +42,9 @@
             builder.RegisterType<CartonProxy>().As<ICartonRepository>().WithParameter("rootUri", ConfigurationManager.Configuration["PROXY_ROOT"]);
             builder.RegisterType<ProductRangeProxy>().As<IProductRangeRepository>().WithParameter("rootUri", ConfigurationManager.Configuration["PROXY_ROOT"]);
             builder.RegisterType<SalesProductProxy>().As<ISalesProductRepository>().WithParameter("rootUri", ConfigurationManager.Configuration["PROXY_ROOT"]);
+
+            // Oracle proxies
+            builder.RegisterType<StockTriggerLevelReportProxy>().As<IStockTriggerLevelReportService>();
         }
     }
 }
