+++ src/Proxy/Proxy.csproj
@@ -8,6 +8,7 @@
 
   <ItemGroup>
     <PackageReference Include="Linn.Common.Proxy" Version="3.0.1" />
+    <PackageReference Include="Oracle.ManagedDataAccess.Core" Version="2.18.3" />
   </ItemGroup>
 
   <ItemGroup>
